Changed the function so that it automatically removes pixels with no cells
Based on this change, functions now automatically remove pixels with no cells. Change was made based on the fact that NA values are considered as non-zero elements in dgCMatrix and that fun = "sum" was not able to remove pixels with no cells even when na.rm = TRUE. Based on this change, na.rm argument is removed. Since pixel IDs are set before removing any NAs, this should not interfere with cell type colocalization analysis that requires the intersection of two rasterized datasets.